+++ libft/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: opavliuk <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/22 16:47:46 by opavliuk          #+#    #+#             */
-/*   Updated: 2018/03/22 17:23:55 by opavliuk         ###   ########.fr       */
+/*   Updated: 2018/03/28 20:17:47 by opavliuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strcmp(const char *s1, const char *s2)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0')
