+++ libft/ft_strcpy.c
@@ -6,7 +6,7 @@
 /*   By: opavliuk <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/21 14:55:24 by opavliuk          #+#    #+#             */
-/*   Updated: 2018/03/22 17:21:39 by opavliuk         ###   ########.fr       */
+/*   Updated: 2018/03/28 19:47:04 by opavliuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strcpy(char *dst, const char *src)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] != '\0')
