+++ libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: opavliuk <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/23 16:37:57 by opavliuk          #+#    #+#             */
-/*   Updated: 2018/03/28 15:35:30 by opavliuk         ###   ########.fr       */
+/*   Updated: 2018/03/28 20:20:35 by opavliuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s && (*f))
